refactor(organization): 🧩 extract useRoles and useOrganizationTypes to shared query module
Moved role and organization type hooks from component to `src/queries/organization.ts` to centralize logic and improve reusability.
